Dev Avancement du 10/12 vers Main (#29)
* Main vers dev le last (#2)

* Initial users avancement

* Revert "Initial users avancement"

This reverts commit 5fe25f360dbb690180352b128f5d5bbfa2fece99.

* User login & user register (#1)

* Advance users

* Update voir desc

Update index
Update global
Update package json

* Appart Controller for creating appart

* Modify appart model

Suppression de la tentative de tableau

* package json fix bug

* Add route for creating an appart

* Service for creating appart

* Create Appart repo

* User login & user register Merge avancement vers dev (#6)

* Advance users

* Update voir desc

Update index
Update global
Update package json

* Update fix

* Update mvc

Update controller, repo, user, userService

* Ajouts middlewares log & auth

Auth middleware, authorization & admin

* Update index / global & packages

modif index, global & packages

* Update Register user & git ignore (#9)

* Modification message d'erreur service TO controller

* User system users route Merge vers dev (#13)

* Dev vers Main (Fonctionnalité register user) (#10)

* Main vers dev le last (#2)

* Initial users avancement

* Revert "Initial users avancement"

This reverts commit 5fe25f360dbb690180352b128f5d5bbfa2fece99.

* User login & user register (#1)

* Advance users

* Update voir desc

Update index
Update global
Update package json

* Appart Controller for creating appart

* Modify appart model

Suppression de la tentative de tableau

* package json fix bug

* Add route for creating an appart

* Service for creating appart

* Create Appart repo

* User login & user register Merge avancement vers dev (#6)

* Advance users

* Update voir desc

Update index
Update global
Update package json

* Update fix

* Update mvc

Update controller, repo, user, userService

* Ajouts middlewares log & auth

Auth middleware, authorization & admin

* Update index / global & packages

modif index, global & packages

* Update Register user & git ignore (#9)

---------


* Avancement User System

Ajout route /users
Ajout /users/:id
Fix de quelques trucs

---------


* Delete appart done without authentification

* Authorization system for delete appart

Missing nice error messages

* Avancement Users Search & Delete (#14)

AJout search de users
Delete user
update index

* Book appart ----> Vers Dev (#15)

* Dev vers Main (Fonctionnalité register user) (#10)

* Main vers dev le last (#2)

* Initial users avancement

* Revert "Initial users avancement"

This reverts commit 5fe25f360dbb690180352b128f5d5bbfa2fece99.

* User login & user register (#1)

* Advance users

* Update voir desc

Update index
Update global
Update package json

* Appart Controller for creating appart

* Modify appart model

Suppression de la tentative de tableau

* package json fix bug

* Add route for creating an appart

* Service for creating appart

* Create Appart repo

* User login & user register Merge avancement vers dev (#6)

* Advance users

* Update voir desc

Update index
Update global
Update package json

* Update fix

* Update mvc

Update controller, repo, user, userService

* Ajouts middlewares log & auth

Auth middleware, authorization & admin

* Update index / global & packages

modif index, global & packages

* Update Register user & git ignore (#9)

---------


* Avancement book aparts check books

* verif authentification pour reservation

---------


* delete appart done

* corrige bug

* editAppart avancement 06/12 (#17) -> Vers Dev

Fix de plusieurs codes dans la chaine de l'archi (route /apparts et /apparts/:id)
Ajout du path d'appart (edit d'appart)

* Delete reservation -> vers dev (#18)

* Dev vers Main (Fonctionnalité register user) (#10)

* Main vers dev le last (#2)

* Initial users avancement

* Revert "Initial users avancement"

This reverts commit 5fe25f360dbb690180352b128f5d5bbfa2fece99.

* User login & user register (#1)

* Advance users

* Update voir desc

Update index
Update global
Update package json

* Appart Controller for creating appart

* Modify appart model

Suppression de la tentative de tableau

* package json fix bug

* Add route for creating an appart

* Service for creating appart

* Create Appart repo

* User login & user register Merge avancement vers dev (#6)

* Advance users

* Update voir desc

Update index
Update global
Update package json

* Update fix

* Update mvc

Update controller, repo, user, userService

* Ajouts middlewares log & auth

Auth middleware, authorization & admin

* Update index / global & packages

modif index, global & packages

* Update Register user & git ignore (#9)

---------


* Avancement book aparts check books

* verif authentification pour reservation

* cancel reservation

* reservation canceled

* change

---------


* Check fix 06 12 ------- Vers dev (#19)

* Fix & Update AppartController

* Delete t.js

* Update & fix route reservation

* Update & fix ReservationService

* Update & fix ReservationRepo

* Ajout dateTimeMiddleware

* Update checkIfadmin

* Update & fix reservation controller

* Edit user finish (#20)

* Start editUser

Algos de tri de données pour le repo

* edit UserDone

Repo et messages d'erreurs / validation + redirection

* Edit user -> Dev (#22)

* Start editUser

Algos de tri de données pour le repo

* edit UserDone

Repo et messages d'erreurs / validation + redirection

* Correction bugs secu edit

---------


* add appart verif for delete (#23)

* test modification modifPostAppart (#24)

* Commit Fix 08_12 (#25)

Fix de la chaîne User


* Edit bug of validation appart

* searchAppartBy (#26)

* Traduction Anglais --> Français

Remplacement de appartement par logement car plus commun

* add link to return to redirect users

* change route user/me

* Ajout UserCreate (#28)

---------

+++ src/controller/AppartController.js
@@ -1,4 +1,5 @@
 const AppartService = require("../services/AppartService");
+const baseURL = require("../../config.json").baseUrl
 
 class AppartController{
     constructor(){
@@ -11,7 +12,7 @@ class AppartController{
             res.status(200).json({count: data.length, apparts: data});
         } catch (error) {
             console.log(error);
-            res.status(500).json({error: "Erreur durant la récupération des appartements"});
+            res.status(500).json({error: "Erreur durant la récupération des logements"});
         }
     }
 
@@ -19,36 +20,53 @@ class AppartController{
         try {
             let data = await this.service.getAppartById(req.params.id);
             if (data) res.status(200).json(data);
-            else res.status(500).json({error: "Erreur durant la récupération de l'appartement"});
+            else res.status(500).json({error: "Erreur durant la récupération du logement"});
         } catch (error) {
             console.log(error);
-            res.status(500).json({error: `Erreur durant la récupération de l'appart ${req.params.id}`});
+            res.status(500).json({error: `Erreur durant la récupération du logement ${req.params.id}`});
         }
     }
 
     async postAppart(req, res){
         try {
-            const body = req.body;
-            let result = await this.service.createAppart(body, res);
-            if (result === "ok") res.status(200).json({message: "success"});
+            const title = req.body.title;
+            const owner = req.body.owner;
+            const address = req.body.address;
+            const status = req.body.status;
+            const price = req.body.price;
+            const area = req.body.area;
+            const nb_rooms = req.body.nb_rooms;
+            const max_people = req.body.max_people;
+
+            const userRank = req.user.rank;
+            const userId = req.user.userId;
+
+            let result = await this.service.createAppart(owner, title, address, status, price, area, nb_rooms, max_people, userRank, userId);
+            if (result) res.status(200).json({message: "success", info: {lien: `${baseURL}/apparts/${result[0].id}`,method: "GET"}});
             else res.status(500).json({error: result});
         } catch (error) {
             console.log(error);
-            res.status(500).json({error: "Erreur durant la création de l'appartement"});
+            res.status(500).json({error: "Erreur durant la création du logement"});
         }
     }
     
     async deleteAppart(req, res){
         try {
-            let result = await this.service.deleteAppart(req, res);
-            if (result === "ok") {
-                res.status(200).json({message: "Appartement supprimé"});
-            } else {
-                res.status(500).json({error: result});
+            const appartId = req.params.id;
+            const check = await this.service.getAppartById(appartId);
+            if (check) {
+                let result = await this.service.deleteAppart(req, res);
+                if (result === "ok") {
+                    res.status(200).json({message: "Logement supprimé"});
+                } else {
+                    res.status(403).json({error: result});
+                }
+            }else{
+                res.status(403).json({error: "Ce logement n'existe pas."});
             }
         } catch (error) {
             console.log(error);
-            res.status(500).json({error: `Error during the delete of the appartment ${req.params.id}`});
+            res.status(500).json({error: `Une erreur est survenue lors de la suppression du logement ${req.params.id}`});
         }
     }
 
@@ -60,10 +78,41 @@ class AppartController{
             const isAdmin = (req.user.rank === "admin") ? true : false;
             let resEdit = await this.service.editAppart(userId, appartId, data, isAdmin);
             if (resEdit) res.status(200).json({message: resEdit});
-            else res.status(400).json({message: "Error during request"});
+            else res.status(400).json({message: "Une erreur est survenue durant la requête"});
         } catch (error) {
             console.log(error);
-            res.status(500).json({error: `Error during the patch of the appartment ${req.params.id}`});
+            res.status(500).json({error: `Une erreur est survenue durant la modification du logement${req.params.id}`});
+        }
+    }
+    
+    async validAppart(req, res) {
+        try {
+            const appartId = req.params.id;
+            let resValid = await this.service.validAppart(appartId);
+            if (resValid == "ok") res.status(200).json({message: "Logement validé !"});
+            else res.status(400).json({message: resValid});
+        } catch (error) {
+            console.log(error);
+            res.status(500).json({error: `Une erreur est survenue durant la validation du logement ${req.params.id}`});
+        }
+    }
+
+    async searchAppartBy(req, res){
+        try {
+            let query = req.query;
+            if (Object.keys(query).length) {
+                let apparts = await this.service.searchAppartBy(query);
+                if (apparts) {
+                    res.status(200).json(apparts);
+                }else{
+                    res.status(400).json({message: "Erreur durant la récupération des logements"})
+                }
+            }else{
+                res.status(400).json({message: "Paramètres manquants"});
+            }
+        } catch (error) {
+            console.log(error);
+            res.status(500).json({error: "Erreur durant la recherche d'un appartement."});
         }
     }
 }
